Move NS1 configuration to external-dns (#2042)
Similarly to what we did with AWS Route53 (#1856), we are migrating the configuration of other DNS providers to the upstream helm chart of external dns. This PR does this migration for NS1.

The ns1 configuration is removed from the k8gb helm chart, and the examples are adapted with the equivalent configuration using the upstream external dns helm chart.

k8gb's helm chart also allowed to ignoreSSL, which can be achieved with the following configuration:
```
  extraArgs:
    ns1-ignoressl:
```

For additional options, please refer to: https://kubernetes-sigs.github.io/external-dns/v0.14.0/tutorials/ns1/

---

To check it for yourself, run the following commands on master and on this branch (the configuration is equivalent, only some labels differ):
```
cd chart/k8gb/
helm template . -f ../../docs/examples/ns1/k8gb-cluster-ns1-eu-west-1.yaml --include-crds > after.yaml
```

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>